customer container double
int credit -> double credit

+++ model/CustomerContainer.java
@@ -19,7 +19,7 @@ public class CustomerContainer {
         return instance;
     }
     
-    public Customer createCustomer(String name, String phoneNumber, String address, int discount, String group, int credit){
+    public Customer createCustomer(String name, String phoneNumber, String address, int discount, String group, double credit){
         Customer customer = new Customer(name, phoneNumber, address, discount, group, credit);
         addCustomer(customer, phoneNumber);
         return customer;
